backend/requirements.txt
Flask
psycopg2-binary
psycopg2
Flask-Migrate
Flask-JWT-Extended
python-dotenv
